Update MELPA to build both unstable and stable.
- Combine stable and unstable builds into the same rotation so one
builds after the other.

- Move all the services to docker containers,
- nginx: contains everything for handling http traffic including log
  rotation and ssl certs.
- syncer: manages syncing all the html files to the latest build
  information.
- logprocessor: processes the http logs.
- rsyncd: rsync for melpa repo.
