--- cfg/README.md
@@ -1,12 +1,0 @@
-# MELPA Config Files
-
-This directory contains config files for applications needed to
-maintain the MELPA infrastructure.
-
-Currently this directory contains local config files for *nginx* and
-*logrotate* which are run by the user processes.
-
-The `etc` directory contains system-level configs required to
-configure the main `runit` process to run user services as well as the
-main *nginx* proxy.
-
