--- cfg/etc/runit/melpa/run
@@ -1,3 +1,0 @@
-#!/bin/sh -e
-exec 2>&1 
-exec chpst -u melpa runsvdir /home/melpa/service 'log: ...........................................................................................................................................................................................................................................................................................................................................................................................................'
