--- cfg/logrotate
@@ -1,19 +1,0 @@
-/home/melpa/log/melpa*.log {
-	daily
-	create 0640 melpa melpa
-	compress
-	dateext
-	missingok
-	missingok
-	notifempty
-	rotate 36500
-	sharedscripts
-	prerotate
-		if [ -d /etc/logrotate.d/httpd-prerotate ]; then \
-			run-parts /etc/logrotate.d/httpd-prerotate; \
-		fi; \
-	endscript
-	postrotate
-		[ ! -f /home/melpa/var/run/nginx.pid ] || kill -USR1 `cat /home/melpa/var/run/nginx.pid`
-	endscript
-}
