+++ docker/docker-compose.yml
@@ -1,0 +1,44 @@
+version: "3"
+services:
+  nginx:
+    build: ./nginx
+    image: melpa/nginx:v1
+    restart: unless-stopped
+    volumes:
+      - /mnt/store/melpa:/mnt/store/melpa:ro
+      - /mnt/store/letsencrypt:/etc/letsencrypt
+      - /mnt/store/log:/mnt/store/log
+      - /mnt/store/log-stable:/mnt/store/log-stable
+    ports:
+      - "80:80"
+      - "443:443"
+  builder:
+    build: ./builder
+    image: melpa/builder:v1
+    restart: unless-stopped
+    volumes:
+      - /mnt/store/melpa:/mnt/store/melpa
+  logprocessor:
+    build: ./logprocessor
+    image: melpa/logprocessor:v1
+    restart: unless-stopped
+    volumes:
+      - /mnt/db:/mnt/db
+      - /mnt/store/melpa:/mnt/store/melpa
+      - /mnt/store/log:/mnt/store/log:ro
+      - /mnt/store/log-stable:/mnt/store/log-stable:ro
+  syncer:
+    build: ./syncer
+    image: melpa/syncer:v1
+    restart: unless-stopped
+    volumes:
+      - /mnt/store/melpa:/mnt/store/melpa
+  rsyncd:
+    build: ./rsyncd
+    image: melpa/rsyncd:v1
+    restart: unless-stopped
+    ports:
+      - "873:873"
+    volumes:
+      - /mnt/store/melpa/packages:/packages:ro
+      - /mnt/store/melpa/packages-stable:/packages-stable:ro
