+++ docker/rsyncd/rsyncd.conf
@@ -1,0 +1,15 @@
+max connections = 1
+log file = /var/log/rsync.log
+timeout = 300
+
+[packages]
+comment = MELPA packages
+path = /packages
+read only = yes
+list = yes
+
+[packages-stable]
+comment = MELPA stable packages
+path = /packages-stable
+read only = yes
+list = yes
