+++ html-stable/Makefile
@@ -1,0 +1,1 @@
+../html/Makefile
