+++ html-stable/donate.png
@@ -1,0 +1,1 @@
+../html/donate.png
