+++ html-stable/favicon.ico
@@ -1,0 +1,1 @@
+../html/favicon.ico
