+++ html-stable/index.html
@@ -1,0 +1,1 @@
+../html/index.html
