+++ html-stable/js
@@ -1,0 +1,1 @@
+../html/js
