+++ html-stable/jslicense.html
@@ -1,0 +1,1 @@
+../html/jslicense.html
