+++ html-stable/packages
@@ -1,0 +1,1 @@
+../packages-stable
