+++ html-stable/partials
@@ -1,0 +1,1 @@
+../html/partials
