+++ html-stable/robots.txt
@@ -1,0 +1,1 @@
+../html/robots.txt
