+++ html-stable/updates.rss.erb
@@ -1,0 +1,1 @@
+../html/updates.rss.erb
