+++ packages-stable/index.html
@@ -1,0 +1,1 @@
+../packages/index.html
