--- scripts/bootstrap
@@ -1,39 +1,0 @@
-#!/bin/bash
-
-# \curl -L https://raw.github.com/melpa/melpa/master/scripts/bootstrap | bash
-
-SUDOENV='DEBIAN_FRONTEND=noninteractive'
-
-cd ${HOME}
-#sudo ${SUDOENV} add-apt-repository -y ppa:cassou/emacs
-sudo ${SUDOENV} apt-get -y update
-sudo ${SUDOENV} apt-get -y upgrade
-sudo ${SUDOENV} apt-get -y install \
-    curl \
-    emacs24 emacs24-el emacs24-common-non-dfsg \
-    git \
-    make \
-    mercurial \
-    nginx \
-    runit \
-    texinfo \
-    tmux
-
-# build the log stuff
-mkdir -p ~/log
-mkdir -p ~/var/run
-mkdir -p ~/var/lib/logrotate
-mkdir -p ~/etc
-
-# ruby is used for templating
-\curl -L https://get.rvm.io | bash -s stable
-PATH=$PATH:$HOME/.rvm/bin
-source "$HOME/.rvm/scripts/rvm"
-rvm requirements
-rvm install ruby-1.9.3
-rvm use ruby-1.9.3 --default
-echo "source ~/.profile" >> .bash_profile
-
-git clone http://github.com/melpa/melpa
-
-sudo reboot
