--- service/builder-stable/log/run
@@ -1,3 +1,0 @@
-#!/bin/sh -e
-[ -d main ] || mkdir -p main
-exec svlogd -tt main
