--- service/builder/run
@@ -1,54 +1,0 @@
-#!/bin/bash -e
-
-SCRIPTPATH=$( cd "$(dirname "$0")" ; pwd -P )
-export HOME=$( cd "${SCRIPTPATH}/../../.." ; pwd -P )
-
-MELPA_REPO="${SCRIPTPATH}/../.."
-cd "${MELPA_REPO}"
-MELPA_BRANCH=$( git rev-parse --abbrev-ref HEAD )
-
-BUILD_STATUS_JSON=${MELPA_REPO}/html/build-status.json
-BUILD_LAST_DURATION_FILE=${MELPA_REPO}/html/.last-build-duration
-
-# Amount of time to sleep between rebuilds
-BUILD_DELAY=$((10 * 60))
-
-## update MELPA repo
-git fetch origin
-git reset --hard "origin/${MELPA_BRANCH}"
-git pull origin "${MELPA_BRANCH}"
-echo
-
-if [ -f "$BUILD_LAST_DURATION_FILE" ]; then
-    BUILD_DURATION=$(cat "$BUILD_LAST_DURATION_FILE")
-fi
-
-update_json() {
-    cat <<EOF > $BUILD_STATUS_JSON
-{
-  "started": $BUILD_STARTED,
-  "completed": ${BUILD_COMPLETED-null},
-  "next": ${BUILD_NEXT-null},
-  "duration": ${BUILD_DURATION-null}
-}
-EOF
-    echo "Writing $BUILD_STATUS_JSON"
-    cat "$BUILD_STATUS_JSON"
-}
-
-
-BUILD_STARTED=$(date "+%s")
-update_json
-
-# Build all the packages.
-scripts/parallel_build_all
-
-# Store completed date
-BUILD_COMPLETED=$(date "+%s")
-BUILD_DURATION=$((BUILD_COMPLETED - BUILD_STARTED))
-echo -n "$BUILD_DURATION" > $BUILD_LAST_DURATION_FILE
-BUILD_NEXT=$((BUILD_COMPLETED + BUILD_DELAY))
-update_json
-
-# Sleep for a while before rebuilding.
-sleep $BUILD_DELAY
