--- service/nginx-stable/run
@@ -1,3 +1,0 @@
-#!/bin/bash -e
-
-/usr/bin/authbind /usr/sbin/nginx -g 'error_log /dev/null crit;' -c /home/melpa/service/nginx-stable/config
