--- service/nginx/config
@@ -1,91 +1,0 @@
-worker_processes 4;
-pid /home/melpa/var/run/nginx.pid;
-daemon off;
-
-events {
-	worker_connections 768;
-}
-
-http {
-	sendfile on;
-	tcp_nopush on;
-	tcp_nodelay on;
-	keepalive_timeout 65;
-	types_hash_max_size 2048;
-	server_tokens off;
-
-	# server_names_hash_bucket_size 64;
-
-	charset utf-8;
-	
-	include /etc/nginx/mime.types;
-	default_type application/octet-stream;
-
-	access_log  /home/melpa/log/melpa.access.log combined;
-	error_log  /home/melpa/log/melpa.error.log info;
-
-	chunked_transfer_encoding off;
-	
-	gzip on;
-	gzip_disable "msie6|Emacs";
-	
-	gzip_vary on;
-	gzip_proxied any;
-	gzip_comp_level 6;
-	gzip_buffers 16 8k;
-	gzip_http_version 1.1;
-	gzip_types text/plain text/xhtml text/css text/js text/csv application/javascript application/x-javascript application/json application/xml text/xml application/atom+xml application/rss+xml;
-
-	client_max_body_size       10m;
-	client_body_buffer_size    128k;
-
-	server {
-	        server_name melpa.milkbox.net;
-	        rewrite ^ $scheme://melpa.org$request_uri? permanent;
-	}
-
-	
-	server {
-		listen  80;
-		listen	443 ssl;
-	
-	        server_name localhost melpa.org www.melpa.org;
-
-		root /home/melpa/melpa/html;
-	
-	        ssl_certificate /etc/letsencrypt/live/melpa.org/fullchain.pem;
-	        ssl_certificate_key /etc/letsencrypt/live/melpa.org/privkey.pem;
-	
-		# Lock down ciphers / SSL versions to attain a good security rating
-		# https://www.ssllabs.com/ssltest/analyze.html?d=melpa.org
-                ssl_protocols TLSv1 TLSv1.1 TLSv1.2;
-                ssl_prefer_server_ciphers on;
-                ssl_dhparam /etc/ssl/certs/dhparam.pem;
-                ssl_ciphers 'ECDHE-RSA-AES128-GCM-SHA256:ECDHE-ECDSA-AES128-GCM-SHA256:ECDHE-RSA-AES256-GCM-SHA384:ECDHE-ECDSA-AES256-GCM-SHA384:DHE-RSA-AES128-GCM-SHA256:DHE-DSS-AES128-GCM-SHA256:kEDH+AESGCM:ECDHE-RSA-AES128-SHA256:ECDHE-ECDSA-AES128-SHA256:ECDHE-RSA-AES128-SHA:ECDHE-ECDSA-AES128-SHA:ECDHE-RSA-AES256-SHA384:ECDHE-ECDSA-AES256-SHA384:ECDHE-RSA-AES256-SHA:ECDHE-ECDSA-AES256-SHA:DHE-RSA-AES128-SHA256:DHE-RSA-AES128-SHA:DHE-DSS-AES128-SHA256:DHE-RSA-AES256-SHA256:DHE-DSS-AES256-SHA:DHE-RSA-AES256-SHA:AES128-GCM-SHA256:AES256-GCM-SHA384:AES128-SHA256:AES256-SHA256:AES128-SHA:AES256-SHA:AES:CAMELLIA:DES-CBC3-SHA:!aNULL:!eNULL:!EXPORT:!DES:!RC4:!MD5:!PSK:!aECDH:!EDH-DSS-DES-CBC3-SHA:!EDH-RSA-DES-CBC3-SHA:!KRB5-DES-CBC3-SHA';
-                ssl_session_timeout 1d;
-                ssl_session_cache shared:SSL:50m;
-                ssl_stapling on;
-                ssl_stapling_verify on;
-                add_header Strict-Transport-Security max-age=15768000;
-	
-	        location ~ /.well-known {
-	                allow all;
-			root /usr/share/nginx/html;
-	        }
-	
-		error_page   500 502 503 504  /50x.html;
-		location = /50x.html {
-			root   /var/www/nginx-default;
-		}
-		location = /packages/archive-contents {
-			default_type text/plain;
-		}
-		location ~ ^/packages/.*\.el {
-			default_type text/plain;
-		}
-		location ~ ^/packages/.*\.svg {
-			add_header Cache-Control no-cache;
-		}
-	}
-}
-
