--- service/process_log/run
@@ -1,11 +1,0 @@
-#!/bin/bash -e
-
-SCRIPTPATH=$( cd "$(dirname "$0")" ; pwd -P )
-MELPA_REPO="${SCRIPTPATH}/../.."
-cd "${MELPA_REPO}"
-
-/usr/bin/python ${MELPA_REPO}/scripts/process_log.py
-
-/usr/sbin/logrotate -s ${MELPA_REPO}/../var/lib/logrotate/status ${MELPA_REPO}/cfg/logrotate
-
-sleep 6h
