--- service/reporter/run
@@ -1,24 +1,0 @@
-#!/bin/bash -e
-
-# Service for reporting missing and old packages.
-
-SCRIPTPATH=$( cd "$(dirname "$0")" ; pwd -P )
-MELPA_REPO="${SCRIPTPATH}/../.."
-cd "${MELPA_REPO}"
-
-BODY=`mktemp`
-
-echo "To: MP <dcurtis@milkbox.net>" > ${BODY}
-echo "Subject: [MELPA] `date "+%Y%m%d %H:%M %z"`" >> ${BODY}
-echo >> ${BODY}
-# echo "# Old Packages" >> ${BODY}
-# echo >> ${BODY}
-# scripts/expired >> ${BODY}
-echo >> ${BODY}
-echo "# Missing Packages" >> ${BODY}
-echo >> ${BODY}
-scripts/missing >> ${BODY}
-
-/usr/sbin/sendmail -t < ${BODY}
-
-sleep 12h
