--- service/syncer/run
@@ -1,14 +1,0 @@
-#!/bin/bash -e
-
-SCRIPTPATH=$( cd "$(dirname "$0")" ; pwd -P )
-MELPA_REPO="${SCRIPTPATH}/../.."
-cd "${MELPA_REPO}"
-
-[[ -s "../.rvm/scripts/rvm" ]] && source "../.rvm/scripts/rvm"
-
-make cleanup
-make json
-make html
-
-# Sync every 5 minutes.
-sleep 5m
